requirements.txt
Django==2.1.7
django-model-utils==3.1.2
django-mptt==0.9.1
django-allauth==0.38.0
djangorestframework==3.9.1
django-debug-toolbar==1.11
Pillow==5.4.1
Selenium